fix: flixhq changed upcloud to vidcloud
